release 0.2.0 (#18)
* Adds new project structure.

Changes:
1. We have moved to new project structure: from a single file to a package.
2. We added support for `pytest` without touching the tests itself
3. Added a lot of utility files
4. Refactored `setup.py` for clean working layout
5. Some small style fixes before the `pylint` era
6. Changed `python` versions we support: droped 2.6, added 3.6

* Adds `tox` support

Changes:
1. Adds `tox` support for testing across different python versions
2. Adds fixes to `travis.yml`
3. Updates `README.rst` with testing instructions

* setup.py: url back to github.com/vsudilov

* Updates README.rst with code example (#14)

* Updates README.rst with code example

* Adds some small typo fixes

* travis: deploy on release (#15)

* travis: deploy on release
* travis: add TOX_ENV=cover stage+condition
* travis: define TOX_ENV=cover in tox.ini
* travis: back out nosetests, use py.test with coverage
* tests: explicitly define tests/ dir

* travis: explicit pip install coveralls (#17)

* travis: explicit pip install coveralls

* travis: use coverage instead of pytest-cov
