+++ flask_consulate/__init__.py
@@ -1,0 +1,4 @@
+# coding: utf-8
+
+from flask_consulate.service import ConsulService
+from flask_consulate.consul import Consul
