+++ setup.cfg
@@ -1,0 +1,7 @@
+[aliases]
+test = pytest
+
+[tool:pytest]
+addopts = -v -x
+norecursedirs = build
+testpaths = tests
