+++ setup.py
@@ -5,58 +5,79 @@ flask-consulate
 
 flask extension that provides an interface to consul via a flask.app
 """
-from setuptools import setup
-import os
+
 import sys
-if os.environ.get('USER', '') == 'vagrant':
-    del os.link
+
+from setuptools import setup, find_packages
+
+
+# Dependencies:
 
 if sys.version_info[0] == 2:
-    dnspython = "dnspython"
+    dnspython = 'dnspython'
 elif sys.version_info[0] == 3:
-    dnspython = "dnspython3"
+    dnspython = 'dnspython3'
+else:
+    raise ValueError('Unsupported python version')
+
+INSTALL_REQUIRES = [
+    'consulate',
+    'Flask',
+    'requests>=2.7.0',
+    dnspython,
+]
+
+TESTS_REQUIRE = [
+    'httpretty',
+    'mock',
+    'six',
+    'coveralls',
+    'coverage',
+    'pytest',
+]
+
+SETUP_REQUIRES = [
+    'pytest-runner',
+]
+
 
 setup(
     name='flask-consulate',
-    version="0.1.2",
-    url='http://github.com/vsudilov/flask-consulate/',
+    version='0.2.0',
+    url='http://github.com/vsudilov/flask-consulate',
     license='MIT',
     author='Vladimir Sudilovsky',
     author_email='vsudilovsky@gmail.com',
+
     description='flask extension that provides an interface to consul via a flask app',
     long_description=__doc__,
-    py_modules=['flask_consulate'],
-    # if you would be using a package instead use packages instead
-    # of py_modules:
-    # packages=['flask_sqlite3'],
+
+    packages=find_packages(),
     zip_safe=False,
     include_package_data=True,
     platforms='any',
-    install_requires=[
-        'consulate',
-        'Flask',
-        'requests>=2.7.0',
-        dnspython,
-        'flask-testing',
-        'httpretty',
-        'mock',
-        'nose',
-        'six',
-        'coveralls'
-    ],
+
+    setup_requires=SETUP_REQUIRES,
+    install_requires=INSTALL_REQUIRES,
+    tests_require=TESTS_REQUIRE,
+    extras_require={
+        'test': TESTS_REQUIRE,
+    },
     test_suite='tests',
+
     classifiers=[
+        'Development Status :: 4 - Beta',
         'Environment :: Web Environment',
         'Intended Audience :: Developers',
         'License :: OSI Approved :: MIT License',
         'Operating System :: OS Independent',
         'Programming Language :: Python',
         'Programming Language :: Python :: 2',
-        'Programming Language :: Python :: 2.6',
         'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.3',
         'Programming Language :: Python :: 3.4',
+        'Programming Language :: Python :: 3.5',
+        'Programming Language :: Python :: 3.6',
         'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
         'Topic :: Software Development :: Libraries :: Python Modules'
     ]
