+++ tox.ini
@@ -1,0 +1,9 @@
+[tox]
+envlist = py{27,34,35,36}
+
+[testenv]
+recreate = False
+deps =
+    .[test]
+commands =
+    python setup.py test
